x64: prelu: eliminate implicit narrowing conversions

+++ src/cpu/x64/prelu/jit_prelu_base_kernel.hpp
@@ -29,16 +29,16 @@ class jit_prelu_base_kernel_t : public jit_generator_t {
 public:
     jit_prelu_base_kernel_t(const cpu_isa_t &isa, const int vlen,
             const prelu::bcast &bcast, const memory_desc_wrapper &tensor_md,
-            const size_t number_vmm_single_compute, const char *name);
+            const int number_vmm_single_compute, const char *name);
 
     int simd_w() const noexcept;
     prelu::bcast get_bcast() const noexcept;
 
 protected:
     int reserve_vmm();
-    int get_compute_vmm(size_t base_idx, size_t unroll_group) const;
+    int get_compute_vmm(int base_idx, int unroll_group) const;
 
-    size_t get_number_reserved_vmms() const noexcept;
+    int get_number_reserved_vmms() const noexcept;
 
     const cpu_isa_t isa_;
     const int simd_w_ = 0;
@@ -52,13 +52,13 @@ private:
     virtual bool any_tensor_bf16() const = 0;
     virtual void load_kernel_call_params() = 0;
     virtual void prepare_kernel_const_vars() = 0;
-    virtual void compute_dst(size_t unrolling_factor, bool tail) = 0;
+    virtual void compute_dst(int unrolling_factor, bool tail) = 0;
     virtual void finalize() = 0;
-    size_t calc_unrolling_factor() const noexcept;
-    size_t calc_tail_size(const memory_desc_wrapper &tensor_md) const noexcept;
+    int calc_unrolling_factor() const noexcept;
+    int calc_tail_size(const memory_desc_wrapper &tensor_md) const noexcept;
     const memory_desc_wrapper tensor_md_;
-    const size_t number_vmm_single_compute_ = 0;
-    size_t number_reserved_vmms_ = 0;
+    const int number_vmm_single_compute_ = 0;
+    int number_reserved_vmms_ = 0;
 };
 
 } // namespace x64
